code-examples/telemetry_traces/consumer/requirements.txt
requests
urllib3==1.26.15
asyncio
memphis-py
google-cloud-bigquery[bqstorage,pandas]
pandas
pytz
